Completed User, post and comment package and tested

+++ src/main/java/com/itsaur/internship/comment/CommentService.java
@@ -5,6 +5,7 @@ import com.itsaur.internship.user.UsersStore;
 import io.vertx.core.Future;
 import io.vertx.core.Vertx;
 
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.UUID;
 
@@ -14,27 +15,65 @@ public class CommentService {
     private UsersStore usersStore;
     Vertx vertx;
 
-    public CommentService(Vertx vertx, CommentStore commentStore) {
-        this.vertx = vertx;
+    public CommentService(Vertx vertx, PostStore postStore, UsersStore usersStore,CommentStore commentStore ) {
         this.commentStore = commentStore;
+        this.postStore = postStore;
+        this.usersStore = usersStore;
+        this.vertx = vertx;
     }
 
-   public Future<Void> addComment(String comment, String username, String filename){
-        return Future.all(this.usersStore.findUserByUsername(username),
-                   this.postStore.findPostByFilename(filename))
+    public Future<Void> addComment(UUID userid, UUID postid, String comment){
+        return Future.all(this.usersStore.findUserByUserid(userid),
+                   this.postStore.findPostByPostid(postid))
+                .onFailure(e -> {
+                    e.printStackTrace();
+                })
                 .compose(res -> {
-                    return this.commentStore.insert(new Comment(comment,
-                                                                res.resultAt(0),
-                                                                res.resultAt(1)));
+                    return this.commentStore.insert(new Comment(UUID.randomUUID(),
+                            LocalDateTime.now(), comment, userid, postid));
                 });
    }
 
-   public Future<Void> changeComment(UUID commentid, String comment){
-        return this.commentStore.update(new Comment(commentid, comment));
+   public Future<Void> changeComment(UUID userid, UUID commentid, String comment){
+        return this.usersStore.findUserByUserid(userid)
+                .onFailure(e -> {
+                    System.out.println("findUserByUserid ");
+                    e.printStackTrace();
+                })
+                .compose(q -> {
+                    if (userid.equals(q.getUserid())) {
+                        return this.commentStore.findById(commentid)
+                                .onFailure(e -> {
+                                    System.out.println("findById");
+                                    e.printStackTrace();
+                                })
+                                .compose(w -> {
+                                    return this.commentStore.update(new Comment(w.getCommentid(),w.getCreatedate(),
+                                            LocalDateTime.now(), comment, w.getUserid(), w.getPostid()));
+
+                                });
+                    } else {
+                        return Future.failedFuture(new IllegalArgumentException("userid of Comment entity and given userid are not the same!"));
+                    }
+                });
    }
 
-   public Future<Void> deleteComment(UUID commentid){
-        return this.commentStore.deleteById(commentid);
+   public Future<Void> deleteComment(UUID userid, UUID commentid){
+        return this.usersStore.findUserByUserid(userid)
+                .compose(res -> {
+                    if (userid.equals(res.getUserid())){
+                        return this.commentStore.findById(commentid)
+                                .onFailure(e -> {
+                                    e.printStackTrace();
+                                })
+                                .compose(w -> {
+                                    return this.commentStore.deleteById(commentid);
+                                });
+                    }else {
+                        return Future.failedFuture("");
+                    }
+
+                });
    }
